expose epm flags up to eris contracts

+++ contracts/operate.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"os"
 	"path"
+	"path/filepath"
 
 	"github.com/eris-ltd/eris-cli/chains"
 	"github.com/eris-ltd/eris-cli/data"
@@ -153,12 +154,7 @@ func DefineAppActionService(do *definitions.Do, app *definitions.Contracts) erro
 	linkAppToChain(do, app)
 
 	if app.AppType.Name == "epm" {
-		if do.Verbose {
-			do.Service.Environment = append(do.Service.Environment, "EPM_VERBOSE=true")
-		}
-		if do.Debug {
-			do.Service.Environment = append(do.Service.Environment, "EPM_DEBUG=true")
-		}
+		prepareEpmAction(do, app)
 	}
 
 	// make data container and import do.Path to do.NewName (if exists)
@@ -212,13 +208,36 @@ func CleanUp(do *definitions.Do, app *definitions.Contracts) error {
 		logger.Debugf("No Throwaway Chain to destroy.\n")
 	}
 
-	logger.Debugf("Removing data dir on host =>\t%s\n", path.Join(common.DataContainersPath, do.Service.Name))
-	os.RemoveAll(path.Join(common.DataContainersPath, do.Service.Name))
+	doData := definitions.NowDo()
+	doData.Name = do.Service.Name
+	doData.Operations = do.Operations
+	if do.NewName != "" {
+		doData.Path = do.NewName
+	}
+	loca := path.Join(common.DataContainersPath, doData.Name)
 
-	logger.Debugf("Removing tmp srv contnr =>\t%s\n", do.Operations.SrvContainerName)
+	logger.Debugf("Exporting Results =>\t\t%s:%s\n", loca, do.Path)
+	data.ExportData(doData)
+
+	if app.AppType.Name == "epm" {
+		files, _ := filepath.Glob(filepath.Join(loca, "epm*"))
+		for _, f := range files {
+			dest := filepath.Join(do.Path, filepath.Base(f))
+			logger.Debugf("Moving file =>\t\t\t%s:%s\n", f, dest)
+			common.Copy(f, dest)
+		}
+	}
+
+	if do.RmD {
+		logger.Debugf("Removing data dir on host =>\t%s\n", path.Join(common.DataContainersPath, do.Service.Name))
+		os.RemoveAll(path.Join(common.DataContainersPath, do.Service.Name))
+	}
+
 	if !do.Rm {
+		logger.Debugf("Removing tmp srv contnr =>\t%s\n", do.Operations.SrvContainerName)
 		perform.DockerRemove(do.Service, do.Operations, true, true)
 	}
+
 	return nil
 }
 
@@ -275,4 +294,66 @@ func linkAppToChain(do *definitions.Do, app *definitions.Contracts) {
 	newLink2 := util.ServiceContainersName("keys", do.Operations.ContainerNumber) + ":" + "keys"
 	do.Service.Links = append(do.Service.Links, newLink)
 	do.Service.Links = append(do.Service.Links, newLink2)
+}
+
+func prepareEpmAction(do *definitions.Do, app *definitions.Contracts) {
+	if do.Verbose {
+		do.Service.Environment = append(do.Service.Environment, "EPM_VERBOSE=true")
+	}
+	if do.Debug {
+		do.Service.Environment = append(do.Service.Environment, "EPM_DEBUG=true")
+	}
+
+	if do.CSV != "" {
+		logger.Debugf("Setting output format to =>\t%s\n", do.CSV)
+		do.Service.EntryPoint = do.Service.EntryPoint + " --output " + do.CSV
+	} else {
+		do.Service.EntryPoint = do.Service.EntryPoint + " --output json"
+	}
+
+	if do.EPMConfigFile != "" {
+		logger.Debugf("Setting config file to =>\t%s\n", do.EPMConfigFile)
+		do.Service.EntryPoint = do.Service.EntryPoint + " --file " + do.EPMConfigFile
+	}
+
+	if len(do.ConfigOpts) != 0 {
+		var toAdd string
+		logger.Debugf("Setting sets file to =>\t%v\n", do.ConfigOpts)
+		for _, s := range do.ConfigOpts {
+			toAdd = toAdd + "," + s
+		}
+		do.Service.EntryPoint = do.Service.EntryPoint + " --set " + toAdd
+	}
+
+	if do.Running {
+		do.Service.EntryPoint = do.Service.EntryPoint + " --summary "
+	}
+
+	if do.ContractsPath != "" {
+		do.Service.EntryPoint = do.Service.EntryPoint + " --contracts-path " + do.ContractsPath
+	}
+
+	if do.ABIPath != "" {
+		do.Service.EntryPoint = do.Service.EntryPoint + " --abi-path " + do.ABIPath
+	}
+
+	if do.DefaultGas != "" {
+		do.Service.EntryPoint = do.Service.EntryPoint + " --gas " + do.DefaultGas
+	}
+
+	if do.Compiler != "" {
+		do.Service.EntryPoint = do.Service.EntryPoint + " --compiler " + do.Compiler
+	}
+
+	if do.DefaultAddr != "" {
+		do.Service.EntryPoint = do.Service.EntryPoint + " --address " + do.DefaultAddr
+	}
+
+	if do.DefaultFee != "" {
+		do.Service.EntryPoint = do.Service.EntryPoint + " --fee " + do.DefaultFee
+	}
+
+	if do.DefaultAmount != "" {
+		do.Service.EntryPoint = do.Service.EntryPoint + " --amount " + do.DefaultAmount
+	}
 }
